requirements.txt
Django==2.0.5
django-cors-headers==2.1.0
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==2.6.11
psycopg2==2.7.4
pylint==1.9.2
pylint-django==0.11.1
pylint-plugin-utils==0.2.6
requests==2.18.4
SQLAlchemy
uWSGI==2.0.15
black
ipython
isort
pycodestyle
